ui: extract avatar component and add avatar next to username on posts
